added initial page dynamicly
