--- .env
@@ -1,8 +1,0 @@
-NEXT_PUBLIC_SERVER_URL=https://uqf4fd0mdidd.usemoralis.com:2053/server
-NEXT_PUBLIC_APP_ID=wsJzCPo9dNXW8FLSwiC3YJO1o4ZvWkeAWs9xHxPt
-
-NEXT_PUBLIC_NG_REDIRECT_URI=http://localhost:3000
-NEXT_PUBLIC_NG_CLIENT_ID=mheQ9lFzaeD1byeY6q9B1VNdVtByq8aDlg05NtDf
-NEXT_PUBLIC_NG_CLIEND_SECRET=BM3xIxky26jpM6M4mPkBNkuVsQWVNMbUo09zCQg8Kl0CYZlt6PTFP7um8Jqb9Lna9LYWGL5iG5b3lD8AIiJqPX0FhsISxKyAMKXqiNpY5JA6Mp05XxE7cZKo2a80qnYZ
-NEXT_PUBLIC_NG_STATE=
-NEXT_PUBLIC_NG_USE_NIFTIES_MOCK=true
