+++ .env.example
@@ -20,3 +20,5 @@ NEXT_PUBLIC_NG_ALONE_TOKEN_ID=
 NEXT_PUBLIC_ORIGINAL_PDF=
 NEXT_PUBLIC_GOLD_PDF=
 NEXT_PUBLIC_DIAMOND_PDF=
+
+NEXT_PUBLIC_INITIAL_PAGE="/issue1" or "/five-stars"
