+++ src/components/common/full-page-loader/full-page-loader.tsx
@@ -1,0 +1,11 @@
+import { Flex, Spinner } from "@chakra-ui/react";
+
+const FullPageLoader = () => {
+  return (
+    <Flex w="100vw" h="100vh" alignItems="center" justifyContent="center">
+      <Spinner color="brand.red" size="xl" />
+    </Flex>
+  );
+};
+
+export { FullPageLoader };
