+++ src/components/common/full-page-loader/index.ts
@@ -1,0 +1,1 @@
+export * from "./full-page-loader";
