--- src/layout/landing/index.ts
@@ -1,1 +1,0 @@
-export * from "./landing";
